enable filters to be used on all items
